backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
openai==1.6.1
python-dotenv==1.0.0
Pillow==10.2.0
pandas==2.1.4
werkzeug==3.0.1
